requirements.txt
openai
langchain
pinecone-client
python-dotenv
PyMuPDF
streamlit
tqdm